pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "MechanicalMind-Dependency-AI"
version = "3.0.1"
authors = [
    {name = "MechMind Team", email = "dev@mechmind.example"}
]
description = "Dependency analysis system for mechanical projects"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.25.1",
    "pyyaml>=6.0",
    "sqlalchemy>=1.4.0",
    "click>=8.0.0",
    "django>=4.0,<5.0"
]

[project.optional-dependencies]
dev = ["black>=25.0", "pre-commit>=3.0"]

[tool.black]
line-length = 88
target-version = ['py310']
